Added new file: cryptoBT.js. Cleaner and more refined alogrithm.
cryptoBT is the reboot of this project. It is a file that contains a
JavaScript object which could be imported into any HTML file and used as
an encryption library to encrypt and decrypt data using a
programmer-defined key and cipher algorithm.

The old files were experimental, and contained flaws in their execution
and implementation. cryptoBT corrects these issues and essentially
accomplishes the original goal: to create a basic encryption library for
personal use on web pages.

Obviously, the data encrypted using this library will not be hidden from
capable malicious users, as JavaScript source is shown in web pages
(though one may find this useful in a Node.js project), however, some
may find this library useful for personal use, educational use, or to
analyze the algorithm and port it to a sever-side language such as PHP,
Perl, ASP.NET, etc. (which I might do in the future.)